Consolodated reusable code into pacakge command service. Separated Link commands into new files.
