+++ src/gluon/commands/packages/LinkExistingApplication.ts
@@ -6,7 +6,6 @@ import {
     MappedParameter,
     MappedParameters,
     Parameter,
-    success,
 } from "@atomist/automation-client";
 import * as _ from "lodash";
 import {QMConfig} from "../../../config/QMConfig";
@@ -19,24 +18,20 @@ import {
     ApplicationService,
     ApplicationType,
 } from "../../util/packages/Applications";
+import {PackageCommandService} from "../../util/packages/PackageCommandService";
 import {
     menuForProjects,
     ProjectService,
 } from "../../util/project/ProjectService";
-import {
-    handleQMError,
-    QMError,
-    ResponderMessageClient,
-} from "../../util/shared/Error";
-import {isSuccessCode} from "../../util/shared/Http";
+import {handleQMError, ResponderMessageClient} from "../../util/shared/Error";
 import {
     RecursiveParameter,
     RecursiveParameterRequestCommand,
 } from "../../util/shared/RecursiveParameterRequestCommand";
 import {menuForTeams, TeamService} from "../../util/team/TeamService";
 
-@CommandHandler("Link an existing library", QMConfig.subatomic.commandPrefix + " link library")
-export class LinkExistingLibrary extends RecursiveParameterRequestCommand {
+@CommandHandler("Link an existing application", QMConfig.subatomic.commandPrefix + " link application")
+export class LinkExistingApplication extends RecursiveParameterRequestCommand {
 
     @MappedParameter(MappedParameters.SlackUserName)
     public screenName: string;
@@ -45,19 +40,19 @@ export class LinkExistingLibrary extends RecursiveParameterRequestCommand {
     public teamChannel: string;
 
     @Parameter({
-        description: "library name",
+        description: "application name",
     })
     public name: string;
 
     @Parameter({
-        description: "library description",
+        description: "application description",
     })
     public description: string;
 
     @Parameter({
         description: "team name",
-        displayable: false,
         required: false,
+        displayable: false,
     })
     public teamName: string;
 
@@ -75,22 +70,27 @@ export class LinkExistingLibrary extends RecursiveParameterRequestCommand {
                 private projectService = new ProjectService(),
                 private memberService = new MemberService(),
                 private bitbucketService = new BitbucketService(),
-                private applicationService = new ApplicationService()) {
+                private applicationService = new ApplicationService(),
+                private packageCommandService = new PackageCommandService()) {
         super();
     }
 
     protected async runCommand(ctx: HandlerContext): Promise<HandlerResult> {
-        try {
-            await ctx.messageClient.addressChannels({
-                text: "🚀 Your new library is being created...",
-            }, this.teamChannel);
 
-            return await this.linkLibraryForGluonProject(
+        logger.debug(`Linking to Gluon project: ${this.projectName}`);
+
+        try {
+            await ctx.messageClient.respond({
+                text: "🚀 Your new application is being created...",
+            });
+
+            return await this.packageCommandService.linkBitbucketRepoToGluonPackage(
                 this.screenName,
                 this.name,
                 this.description,
                 this.bitbucketRepositorySlug,
                 this.projectName,
+                ApplicationType.DEPLOYABLE,
             );
         } catch (error) {
             return await handleQMError(new ResponderMessageClient(ctx), error);
@@ -124,7 +124,7 @@ export class LinkExistingLibrary extends RecursiveParameterRequestCommand {
         if (_.isEmpty(this.bitbucketRepositorySlug)) {
             const project = await this.projectService.gluonProjectFromProjectName(this.projectName);
             if (_.isEmpty(project.bitbucketProject)) {
-                throw new QMError(`The selected project does not have an associated bitbucket project. Please first associate a bitbucket project using the \`${QMConfig.subatomic.commandPrefix} link bitbucket project\` command.`);
+                return await ctx.messageClient.respond(`❗The selected project does not have an associated bitbucket project. Please first associate a bitbucket project using the \`${QMConfig.subatomic.commandPrefix} link bitbucket project\` command.`);
             }
 
             const bitbucketRepos = await this.bitbucketService.bitbucketRepositoriesForProjectKey(project.bitbucketProject.key);
@@ -141,67 +141,4 @@ export class LinkExistingLibrary extends RecursiveParameterRequestCommand {
             );
         }
     }
-
-    private async linkLibraryForGluonProject(slackScreeName: string,
-                                             libraryName: string,
-                                             libraryDescription: string,
-                                             bitbucketRepositorySlug: string,
-                                             gluonProjectName: string): Promise<HandlerResult> {
-        const project = await this.projectService.gluonProjectFromProjectName(gluonProjectName);
-        logger.debug(`Linking Bitbucket repository: ${bitbucketRepositorySlug}`);
-
-        return await this.linkBitbucketRepository(
-            slackScreeName,
-            libraryName,
-            libraryDescription,
-            bitbucketRepositorySlug,
-            project.bitbucketProject.key,
-            project.projectId);
-    }
-
-    private async linkBitbucketRepository(slackScreeName: string,
-                                          libraryName: string,
-                                          libraryDescription: string,
-                                          bitbucketRepositorySlug: string,
-                                          bitbucketProjectKey: string,
-                                          gluonProjectId: string): Promise<HandlerResult> {
-        const repoResult = await this.bitbucketService.bitbucketRepositoryForSlug(bitbucketProjectKey, bitbucketRepositorySlug);
-
-        if (!isSuccessCode(repoResult.status)) {
-            throw new QMError("Unable to find the specified repository in Bitbucket. Please make sure it exists.");
-        }
-
-        const repo = repoResult.data;
-
-        const member = await this.memberService.gluonMemberFromScreenName(slackScreeName);
-
-        const remoteUrl = _.find(repo.links.clone, clone => {
-            return (clone as any).name === "ssh";
-        }) as any;
-
-        const createApplicationResult = await this.applicationService.createGluonApplication(
-            {
-                name: libraryName,
-                description: libraryDescription,
-                applicationType: ApplicationType.LIBRARY,
-                projectId: gluonProjectId,
-                createdBy: member.memberId,
-                bitbucketRepository: {
-                    bitbucketId: repo.id,
-                    name: repo.name,
-                    slug: bitbucketRepositorySlug,
-                    remoteUrl: remoteUrl.href,
-                    repoUrl: repo.links.self[0].href,
-                },
-                requestConfiguration: true,
-            });
-
-        if (!isSuccessCode(createApplicationResult.status)) {
-            logger.error(`Failed to link package. Error: ${JSON.stringify(createApplicationResult)}`);
-            throw new QMError("Failed to link the specified package from bitbucket.");
-        }
-
-        return await success();
-    }
-
 }
